Make function that calculates sales tax based on money spent and rate.

+++ src/thoughtworkschallenge/ParsedItems.java
@@ -1,11 +1,13 @@
 package thoughtworkschallenge;
 
+import java.math.BigDecimal;
+
 public class ParsedItems {
     final Float numberBought;
     final String itemBought;
-    final Float price;
+    final BigDecimal price;
 
-    public ParsedItems(Float numberBought, String itemBought, Float price) {
+    public ParsedItems(Float numberBought, String itemBought, BigDecimal price) {
         this.numberBought = numberBought;
         this.itemBought = itemBought;
         this.price = price;
